update db schema, rewrite dao, add localization resources

+++ src/main/java/com/epam/rd/java/basic/finaltask/selectioncommittee/db/dao/FacultyDAO.java
@@ -1,15 +1,15 @@
 package com.epam.rd.java.basic.finaltask.selectioncommittee.db.dao;
 
-import com.epam.rd.java.basic.finaltask.selectioncommittee.db.entity.Faculty;
+import com.epam.rd.java.basic.finaltask.selectioncommittee.db.entity.impl.Faculty;
 import java.util.List;
 
 public interface FacultyDAO {
 
   int insertFaculty(Faculty faculty);
 
-  boolean deleteFaculty(long id);
+  boolean deleteFaculty(int id);
 
-  Faculty findFacultyById(long id);
+  Faculty findFacultyById(int id);
 
   boolean updateFaculty(Faculty faculty);
